splade/requirements.txt
datasets==4.5.0
python-dotenv==1.1.0
ipykernel==7.2.0
pandas==2.2.3
psycopg2-binary==2.9.10
torch>=2.0.0
transformers>=4.30.0
